+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - groupBy function
 - includes function
 - last function
+- GitHub Actions for CI
+- pluck function
 
 ### Changed
 - added more types